+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # WIP
 * i: Minimum version Chrome/Chromium 146
+* fix: UserScript messages with Firefox (still some issues, but works better)
 
 # 3.7.1
 * i: Minimum version Firefox 148